test Familia finalizados

+++ src/main/java/restproductos/repository/IFamiliaRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import restproductos.modelo.entities.Familia;
 
-public interface IFamiliaRepository extends JpaRepository<Familia, Long> {
+public interface IFamiliaRepository extends JpaRepository<Familia, Integer> {
     // Familia findByNombre(String nombre);
 
 }
